xenapi: Enhance devstack progress monitoring
Fixes bug 1091299

XenServer - devstack install monitors a log file by tailing it, to see,
if the devstack installation is finished. In some cases this script does
not detect, that the startup script is finished, and just waiting,
causing build failures with timeouts. With this change, the
install_os_domU script monitors, if the run.sh script is still running,
thus guaranteed to exit as soon as run.sh is done.

Change-Id: I24a7a46e93ce26be024096828c7954bc694af2b2